saas/requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
psycopg2-binary==2.9.9
gunicorn==21.2.0
redis==5.0.1
celery==5.3.6
stripe==7.0.0
requests==2.31.0
numpy==1.26.4
PyYAML==6.0.1
